Move Iceberg catalog classes into relevant packages
